requirements.txt

httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
orjson>=3.9.15
tenacity>=8.2.3
pytz>=2024.1
